Trying to complete challenge specifications
Trying to fix the coding so it fits the challenge specification. Struggling a bit with having the items spread in a grid style instead of rows. Therefore, so if 4 was input it is 2 x 2 instead of a 4 horizontally. Aiming to complete this with as minimal help or searching on google.